[fix] third batch of CI fix (#2680)

+++ sw/airborne/modules/ctrl/optical_flow_hover.c
@@ -216,8 +216,8 @@ uint8_t hover_method;
 
 /// Function definitions
 // Callback function of the optical flow estimate:
-void ofh_optical_flow_cb(uint8_t sender_id __attribute__((unused)), uint32_t stamp, int16_t flow_x, int16_t flow_y,
-                         int16_t flow_der_x, int16_t flow_der_y, float quality, float size_div);
+void ofh_optical_flow_cb(uint8_t sender_id __attribute__((unused)), uint32_t stamp, int32_t flow_x, int32_t flow_y,
+                         int32_t flow_der_x, int32_t flow_der_y, float quality, float size_div);
 
 // resetting all variables to be called for instance when starting up / re-entering module
 static void reset_horizontal_vars(void);
